Refactor shared cache lookup
