+++ src/data/elasticsearch.rs
@@ -3,6 +3,7 @@ mod data_streams;
 mod ilm_explain;
 mod indices_settings;
 mod nodes;
+mod searchable_snapshots_cache_stats;
 mod version;
 
 pub use aliases::*;
@@ -10,4 +11,5 @@ pub use data_streams::*;
 pub use ilm_explain::*;
 pub use indices_settings::*;
 pub use nodes::*;
+pub use searchable_snapshots_cache_stats::*;
 pub use version::*;
